Use "System Events" instead of whatever app is frontmost
Target the "System Events" application instead of whatever app happens to be frontmost at the time. This ensures that the password is being typed in to a system process, and not just any random app. This should alsö improve reliability as the System Events app is explicitly provided by macOS for this exact purpose. 

Alsö, this script is not portable; there's no need to use "env" to launch "osascript". Apple Events literally can only happen on a Mac, and osascript isn't going anywhere.