Use "System Events" instead of whatever app is frontmost
Target the "System Events" application instead of whatever app happens
to be frontmost at the time when prompting for passwords. This ensures
that the password is being typed in to a system process, and not just
any random app. This should also improve reliability as the System
Events app is explicitly provided by macOS for this exact purpose.