Always use System Events
Tell System Events to get out of the way once it's done to ensure that keyboard focus returns to wherever it was before.
This ensures that (1) System Events is running, (2) that it quits after the dialog, (3) that the dialog text is returned even after System Events quits, and (4) that the `osascript` tool returns an error code when "cancel" is selected.

Using the frontmost application for interaction is an endless rabbithole of edge cases and errors. It seems convenient but it is not a good design.

Resolves #32 and #25