Add SUDO_ASKPASS to environment variables
Alsö also, set DISPLAY to something nonsensical to avoid the possibility of setting it to a valid but undesired value. (E.g., if another user is logged in and has `DISPLAY=0`, but the current user doesn’t have XQuartz running, we don’t want to inadvertantly tell this user’s X apps to appear on the other user’s XQuartz. Or, Xvnc &c.)